Adds support for Yaml Schemas
Currently we only support Json Schemas. This is only 1 of the two
standard formats that Swagger supports, the other being YAML. This
change will add support YAML.

Please note that the reason Symfony YAML is locked to ^2.8.2 is due to
versions prior to this treat lines with a "#" in a quoted string
[incorrectly as a comment][ref]. Swagger makes heavy use of "#" symbols
in it's strings to identify references responses and other things,
making it impossible for us to parse the swagger file.

Fixes issue #21

[ref]: https://github.com/symfony/yaml/commit/5f77b9aeb8f7aeda000bc2fea8874f3ccf1ec523