Follow PSR2
This will make the codebase follow PSR2, and ensure that only compilable
PHP is committed. It ignores and warnings about line length but will
fail the build and tests on any other error.

Having a coding standard lowers the barrier to entry for people coming
to use a project, and PSR-2 is the standard which has been adopted most
widely within the PHP community.